+++ R/raw.fixInvalidFile.R
@@ -1,18 +1,30 @@
-#' Fix invalid filename
+#' Fix invalid data filename
 #'
 #' @param filename name of file
-#' @param instrument new name of instrument
+#' @param addProject project to be added
+#' @param addInstrument instrument to be added
+#' @param addUser user to be added
+#' @param addSample sample to be added
 #' @return returns fixed filenames
 #' @examples
-#' f = raw.findFiles(raw.getSamplePath(), instrument='MM160622SI1')
+#' f = raw.findFiles(raw.getSamplePath(), addInstrument='MM160622SI1')
 #' print(f)
-#' raw.fixInvalidFile(f, instrument='vsm')
+#' f.new = raw.fixInvalidFile(f, instrument='vsm')
+#' print(paste("Fixing:",f,"to new file:",f.new))
 #' @export
-raw.fixInvalidFile <- function(filename, instrument='') {
+raw.fixInvalidFile <- function(filename, addInstrument='', addUser='',
+                               addProject='', addSample='') {
   p = dirname(filename)
   f = basename(filename)
-  if (nchar(instrument)>0) {
-    f = gsub('^(.*?_.*?_.*?)_(.*)$','\\1_vsm_\\2',f)
-  }
+  fs = strsplit(f, '_')
+  if (nchar(addProject)>0)
+    fs = lapply(fs, function(x) { append(x, addProject, after = 1) } )
+  if (nchar(addUser)>0)
+    fs = lapply(fs, function(x) { append(x, addUser, after = 2) } )
+  if (nchar(addInstrument)>0)
+    fs = lapply(fs, function(x) { append(x, addInstrument, after = 3) } )
+  if (nchar(addSample)>0)
+    fs = lapply(fs, function(x) { append(x, addSample, after = 4) } )
+  f = sapply(fs, function(x) { paste(x,collapse = '_') })
   file.path(p,f)
 }